requirements.txt
common==0.1.2
components==0.0.1a0
constant==0.0.4
faiss_cpu==1.10.0
karateclub==1.3.3
networkx==2.6.3
nltk==3.9.1
numpy==1.24.4
openai==1.78.1
pandas==1.5.3
py2neo==2021.2.4
rank_bm25==0.2.2
scikit_learn==1.6.1
sentence_transformers==3.4.1
tqdm==4.66.1
transformers==4.49.0
